Add support for n args for to do and to done.
Click only supports `nargs=-1`, that is an unspecified number of
arguments, for one argument per command. Thus, the task command has been
given priority. This has no impact on `to done`. For `to do`, this means
that rank is specified first, followed by as many tasks as desired;
however, all tasks will be added at the same rank.